Changed Task struct to Job struct
